Review fixes: move TickIcon to ui and unify its API with other icons
