+++ apps/store/src/components/CheckoutPage/CheckoutPage.tsx
@@ -5,7 +5,7 @@ import { useTranslation } from 'next-i18next'
 import Link from 'next/link'
 import { useRouter } from 'next/router'
 import { PropsWithChildren, useState } from 'react'
-import { BankIdIcon, Button, Heading, mq, Space, Text, theme } from 'ui'
+import { BankIdIcon, Button, Heading, mq, Space, Text, TickIcon, theme } from 'ui'
 import { CampaignsSection } from '@/components/CartInventory/CampaignsSection'
 import { CartEntryItem } from '@/components/CartInventory/CartEntryItem'
 import { CartEntryList } from '@/components/CartInventory/CartEntryList'
@@ -26,7 +26,6 @@ import { PageLink } from '@/utils/PageLink'
 import { CartCollapsible } from './CartCollapsible/CartCollapsible'
 import { FormElement } from './CheckoutPage.constants'
 import { CheckoutPageProps } from './CheckoutPage.types'
-import { TickIcon } from './TickIcon'
 import { useHandleSubmitCheckout } from './useHandleSubmitCheckout'
 
 const CheckoutPage = (props: CheckoutPageProps) => {
