+++ packages/ui/src/components/InputBase.tsx
@@ -1,5 +1,5 @@
 import styled from '@emotion/styled'
-import { WarningTriangleIcon } from '../icons/WarningTriangleIcon'
+import { WarningTriangleIcon } from '../icons'
 import { Space } from './Space'
 
 const Label = styled.label(({ theme }) => ({
