+++ packages/ui/src/icons/index.tsx
@@ -1,13 +1,15 @@
-export { MailIcon } from './MailIcon'
-export { PhoneIcon } from './PhoneIcon'
-export { ArrowForwardIcon } from './ArrowIcon'
-export { CrossIcon } from './CrossIcon'
-export { ChevronIcon } from './Chevron'
-export { CheckIcon } from './CheckIcon'
-export { PlayIcon } from './PlayIcon'
-export { PauseIcon } from './PauseIcon'
-export { PlusIcon } from './PlusIcon'
-export { MinusIcon } from './MinusIcon'
-export { AppleIcon } from './AppleIcon'
 export { AndroidIcon } from './AndroidIcon'
+export { AppleIcon } from './AppleIcon'
+export { ArrowForwardIcon } from './ArrowIcon'
 export { BankIdIcon } from './BankIdIcon'
+export { CheckIcon } from './CheckIcon'
+export { ChevronIcon } from './Chevron'
+export { CrossIcon } from './CrossIcon'
+export { MailIcon } from './MailIcon'
+export { MinusIcon } from './MinusIcon'
+export { PauseIcon } from './PauseIcon'
+export { PlayIcon } from './PlayIcon'
+export { PlusIcon } from './PlusIcon'
+export { PhoneIcon } from './PhoneIcon'
+export { TickIcon } from './TickIcon'
+export { WarningTriangleIcon } from './WarningTriangleIcon'
